final_project/requirements.txt
pytest==7.2.0
pytest-xdist==3.1.0
selenium==4.7.2
webdriver-manager==3.8.5
allure-pytest==2.12.0
requests==2.28.1
Faker==15.3.4
sqlalchemy
cryptography
pymysql
fastapi
uvicorn